General cleanup of the nano-sync crate and preparations to be integrated in the nano-blockchain.
