+++ nano-sync/Cargo.toml
@@ -3,35 +3,31 @@ name = "nimiq-nano-sync"
 version = "0.1.0"
 authors = ["The Nimiq Core Development Team <info@nimiq.com>"]
 description = "Recursive SNARKs for Nimiq's Nano Nodes"
-license = "MIT/Apache-2.0"
+license = "Apache-2.0"
 homepage = "https://nimiq.com"
 repository = "https://github.com/nimiq/core-rs-albatross"
 edition = "2018"
 
 [dependencies]
-blake2-rfc = "0.2"
-byteorder = "1.3.4"
-colored = { version = "1.9", optional = true }
-hex = "0.4"
-rand = "0.7"
 thiserror = "1.0.23"
+rand = { version = "0.7", optional = true }
 
-ark-std = { git = "https://github.com/arkworks-rs/utils" }
+ark-std = { git = "https://github.com/arkworks-rs/utils", optional = true }
 ark-ff = { git = "https://github.com/arkworks-rs/algebra" }
 ark-serialize = { git = "https://github.com/arkworks-rs/algebra" }
 ark-ec = { git = "https://github.com/arkworks-rs/algebra" }
 ark-relations = { git = "https://github.com/arkworks-rs/snark" }
-ark-mnt4-753 = { git = "https://github.com/arkworks-rs/curves", features = ["r1cs"] }
-ark-mnt6-753 = { git = "https://github.com/arkworks-rs/curves", features = ["r1cs"] }
-ark-crypto-primitives = { git = "https://github.com/arkworks-rs/crypto-primitives", branch = "main", features = ["r1cs"] }
+ark-mnt4-753 = { git = "https://github.com/arkworks-rs/curves" }
+ark-mnt6-753 = { git = "https://github.com/arkworks-rs/curves" }
+ark-crypto-primitives = { git = "https://github.com/arkworks-rs/crypto-primitives", branch = "main" }
 ark-r1cs-std = { git = "https://github.com/arkworks-rs/r1cs-std" }
-ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", features = ["r1cs"] }
+ark-groth16 = { git = "https://github.com/arkworks-rs/groth16" }
 
 nimiq-bls = { path = "../bls", version = "0.1" }
 nimiq-nano-primitives = { path = "../nano-primitives", version = "0.1"}
 nimiq-primitives = { path = "../primitives", version = "0.1", features = ["policy"] }
 
 [features]
-default = []
-cost-analysis = ["colored"]
-print-trace = ["ark-groth16/print-trace"]
+default = ["prover"]
+verifier = []
+prover = ["ark-std", "rand", "ark-crypto-primitives/r1cs", "ark-mnt4-753/r1cs", "ark-mnt6-753/r1cs", "ark-groth16/r1cs"]
