+++ nano-sync/src/circuits/mnt6/merger_wrapper.rs
@@ -9,7 +9,6 @@ use ark_r1cs_std::prelude::{AllocVar, Boolean, EqGadget};
 use ark_relations::r1cs::{ConstraintSynthesizer, ConstraintSystemRef, SynthesisError};
 
 use crate::utils::{prepare_inputs, unpack_inputs};
-use crate::{end_cost_analysis, next_cost_analysis, start_cost_analysis};
 
 /// This is the merger wrapper circuit. It takes as inputs an initial state commitment, a final state
 /// commitment and a verifying key commitment and it produces a proof that there exists a valid SNARK
@@ -59,21 +58,14 @@ impl ConstraintSynthesizer<MNT6Fr> for MergerWrapperCircuit {
     /// This function generates the constraints for the circuit.
     fn generate_constraints(self, cs: ConstraintSystemRef<MNT6Fr>) -> Result<(), SynthesisError> {
         // Allocate all the constants.
-        #[allow(unused_mut)]
-        let mut cost = start_cost_analysis!(cs, || "Alloc constants");
-
         let vk_merger_var =
             VerifyingKeyVar::<MNT4_753, PairingVar>::new_constant(cs.clone(), &self.vk_merger)?;
 
         // Allocate all the witnesses.
-        next_cost_analysis!(cs, cost, || { "Alloc witnesses" });
-
         let proof_var =
             ProofVar::<MNT4_753, PairingVar>::new_witness(cs.clone(), || Ok(&self.proof))?;
 
         // Allocate all the inputs.
-        next_cost_analysis!(cs, cost, || { "Alloc inputs" });
-
         let initial_state_commitment_var =
             Vec::<FqVar>::new_input(cs.clone(), || Ok(&self.initial_state_commitment[..]))?;
 
@@ -83,8 +75,6 @@ impl ConstraintSynthesizer<MNT6Fr> for MergerWrapperCircuit {
         let vk_commitment_var = Vec::<FqVar>::new_input(cs, || Ok(&self.vk_commitment[..]))?;
 
         // Unpack the inputs by converting them from field elements to bits and truncating appropriately.
-        next_cost_analysis!(cs, cost, || { "Unpack inputs" });
-
         let initial_state_commitment_bits =
             unpack_inputs(initial_state_commitment_var)?[..760].to_vec();
 
@@ -94,8 +84,6 @@ impl ConstraintSynthesizer<MNT6Fr> for MergerWrapperCircuit {
         let vk_commitment_bits = unpack_inputs(vk_commitment_var)?[..760].to_vec();
 
         // Verify the ZK proof.
-        next_cost_analysis!(cs, cost, || { "Verify ZK proof" });
-
         let mut proof_inputs = prepare_inputs(initial_state_commitment_bits);
 
         proof_inputs.append(&mut prepare_inputs(final_state_commitment_bits));
@@ -110,8 +98,6 @@ impl ConstraintSynthesizer<MNT6Fr> for MergerWrapperCircuit {
             &proof_var,
         )?
         .enforce_equal(&Boolean::constant(true))?;
-
-        end_cost_analysis!(cs, cost);
 
         Ok(())
     }
